Add long suffix to moments

+++ ImageObject.cpp
@@ -1,7 +1,8 @@
 #include "ImageObject.h"
 
-ImageObject::ImageObject(const int area, const int perimeter, const int momentX, const int momentY, const int momentX2,
-                         const int momentY2, const int moment11) {
+ImageObject::ImageObject(
+		const long int area, const long int perimeter, const long int momentX, const long int momentY, const long int momentX2,
+		const long int momentY2, const long int moment11) {
 	this->perimeter = perimeter;
 	this->area = area;
 	this->xt = momentX / area;
@@ -10,6 +11,6 @@ ImageObject::ImageObject(const int area, const int perimeter, const int momentX,
 	this->miMin = this->miMax = 0.5 * (momentX2 + momentY2);
 	this->miMin -= 0.5 * std::sqrt(4.0 * moment11 * moment11 + (momentX2 - momentY2) * (momentX2 - momentY2));
 	this->miMax += 0.5 * std::sqrt(4.0 * moment11 * moment11 + (momentX2 - momentY2) * (momentX2 - momentY2));
-	this->F2 = static_cast<double>(miMin / miMax);
+	this->F2 = miMin / miMax;
 
 }
